Fixed samples

+++ samples/Tiny.cpp
@@ -42,8 +42,8 @@ void main() {
 )";
 
 GLFWwindow* window_ = nullptr;
-uint32_t width_ = 800;
-uint32_t height_ = 600;
+int width_ = 800;
+int height_ = 600;
 FramesPerSecondCounter fps_;
 
 std::unique_ptr<lvk::IDevice> device_;
@@ -120,7 +120,7 @@ void VulkanObjects::render() {
   {
     buffer.cmdBindRenderPipeline(renderPipelineState_Triangle_);
     buffer.cmdBindViewport({0.0f, 0.0f, (float)width_, (float)height_, 0.0f, +1.0f});
-    buffer.cmdBindScissorRect({0, 0, width_, height_});
+    buffer.cmdBindScissorRect({0, 0, (uint32_t)width_, (uint32_t)height_});
     buffer.cmdPushDebugGroupLabel("Render Triangle", lvk::Color(1, 0, 0));
     buffer.cmdDraw(lvk::Primitive_Triangle, 0, 3);
     buffer.cmdPopDebugGroupLabel();
